% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{assemble_D_delta_gamma_fn}
\alias{assemble_D_delta_gamma_fn}
\title{Title}
\usage{
assemble_D_delta_gamma_fn(dgamma_coefs, D_bi_set)
}
\arguments{
\item{dgamma_coefs}{coefs}

\item{D_bi_set}{D_bi sets}
}
\value{
a function (u,v) to 2x2 matrix
}
\description{
Title
}
\examples{
assemble_D_delta_gamma_fn(dgamma_coefs,D_bi_set)
}
