% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis_functions.R
\name{build_basis_set}
\alias{build_basis_set}
\title{Title}
\usage{
build_basis_set(Pmax, Qmax)
}
\arguments{
\item{Pmax}{}

\item{Qmax}{}
}
\description{
Title
}
\examples{
basis_set <- build_basis_set(3,3)
}
