% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis_functions.R
\name{build_bi_set}
\alias{build_bi_set}
\title{Title}
\usage{
build_bi_set(basis_set)
}
\arguments{
\item{basis_set}{}
}
\value{
a list of bi set (grad and normalized of scalar basis)
}
\description{
Title
}
\examples{
bi_set <- build_bi_set(basis_set)
}
