% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis_functions.R
\name{build_D_bi_set}
\alias{build_D_bi_set}
\title{Title}
\usage{
build_D_bi_set(basis_set)
}
\arguments{
\item{basis_set}{}
}
\value{
a list of D bi sets
}
\description{
Title
}
\examples{
D_bi_set <- build_D_bi_set(basis_set)
}
