% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surface_functions.R
\name{calc_params}
\alias{calc_params}
\title{Calculate a New Parameterization}
\usage{
calc_params(f, X, init_params)
}
\arguments{
\item{f}{Embedding map.}

\item{X}{Numeric matrix of high-dimensional data.}

\item{t}{Numeric matrix of initial low-dimensional parameterizations.}
}
\value{
A numeric matrix of parameterizations.
}
\description{
Calculate a New Parameterization
}
