% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{grad_basis_cpp}
\alias{grad_basis_cpp}
\title{Gradient of Fourier basis \eqn{\nabla\psi_{pq}(u,v)}.}
\usage{
grad_basis_cpp(u, v, p, q)
}
\arguments{
\item{u}{Numeric input.}

\item{v}{Numeric input.}

\item{p}{Integer Fourier index.}

\item{q}{Integer Fourier index.}
}
\value{
A numeric vector of length 2.
}
\description{
Computes the gradient:
\deqn{
  \nabla\psi_{pq}(u,v) = \left(
    -c_{pq}(2\pi p)\sin(2\pi p u)\cos(2\pi q v),\;
    -c_{pq}(2\pi q)\cos(2\pi p u)\sin(2\pi q v)
  \right).
}
}
