% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{jacobian_grad_cpp}
\alias{jacobian_grad_cpp}
\title{Jacobian of gradient field \eqn{D(\nabla\psi_{pq})(u,v)}.}
\usage{
jacobian_grad_cpp(u, v, p, q)
}
\arguments{
\item{u}{Numeric input.}

\item{v}{Numeric input.}

\item{p}{Integer Fourier index.}

\item{q}{Integer Fourier index.}
}
\value{
A 2×2 numeric matrix.
}
\description{
Computes the 2×2 Jacobian matrix of the gradient field.
}
