% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{jacobian_rot_cpp}
\alias{jacobian_rot_cpp}
\title{Jacobian of rotated gradient field \eqn{D(*\nabla\psi_{pq})(u,v)}.}
\usage{
jacobian_rot_cpp(u, v, p, q)
}
\arguments{
\item{u}{Numeric input.}

\item{v}{Numeric input.}

\item{p}{Integer Fourier index.}

\item{q}{Integer Fourier index.}
}
\value{
A 2×2 numeric matrix.
}
\description{
Computes the 2×2 Jacobian matrix of the rotated gradient.
}
