% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logspace_sum}
\alias{logspace_sum}
\title{Sum in Logspace}
\usage{
logspace_sum(lx, ly)
}
\arguments{
\item{lx}{A double value in logspace}

\item{ly}{A double value in logspace}
}
\value{
A numeric value
}
\description{
Given log(x) and log(y), return log(x+y)
}
