% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rot_basis_cpp}
\alias{rot_basis_cpp}
\title{Rotated gradient \eqn{*\nabla\psi_{pq}(u,v)}.}
\usage{
rot_basis_cpp(u, v, p, q)
}
\arguments{
\item{u}{Numeric input.}

\item{v}{Numeric input.}

\item{p}{Integer Fourier index.}

\item{q}{Integer Fourier index.}
}
\value{
A numeric vector of length 2.
}
\description{
Computes the 90-degree rotated gradient:
\deqn{
  *\nabla\psi_{pq}(u,v) = \left(
    c_{pq}(2\pi q)\cos(2\pi p u)\sin(2\pi q v),\;
   -c_{pq}(2\pi p)\sin(2\pi p u)\cos(2\pi q v)
  \right).
}
}
