% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{scalar_basis_cpp}
\alias{scalar_basis_cpp}
\title{Scalar Fourier basis \eqn{\psi_{pq}(u,v)}.}
\usage{
scalar_basis_cpp(u, v, p, q)
}
\arguments{
\item{u}{Numeric value in \[0,1\].}

\item{v}{Numeric value in \[0,1\].}

\item{p}{Integer Fourier index in the u-direction.}

\item{q}{Integer Fourier index in the v-direction.}
}
\value{
A numeric scalar.
}
\description{
Compute the scalar Fourier basis function
\deqn{\psi_{pq}(u,v) = c_{pq}\cos(2\pi p u)\cos(2\pi q v).}
}
