% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize_gamma_scalar_input}
\alias{visualize_gamma_scalar_input}
\title{Gamma function plot in [0,1]^2}
\usage{
visualize_gamma_scalar_input(
  gamma_fun,
  n = 25,
  arrow_color = "tomato",
  point_color = "grey40",
  warp_scale = 1
)
}
\arguments{
\item{gamma_fun}{the gamma function we want to visualize}

\item{n}{grid number}

\item{arrow_color}{default tomatc}

\item{point_color}{default grey40}

\item{warp_scale}{default 1}
}
\value{
a plot
}
\description{
Gamma function plot in [0,1]^2
}
