% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{assemble_grad_f2k_from_state}
\alias{assemble_grad_f2k_from_state}
\title{Title}
\usage{
assemble_grad_f2k_from_state(state_list, gamma_k, f2_grad_fn, epsilon)
}
\arguments{
\item{state_list}{state_list}

\item{gamma_k}{gamma_k}

\item{f2_grad_fn}{f2_grad_fn}

\item{epsilon}{step to upgrade}
}
\value{
a function (u,v) to 3x2 matrix, named grad_f2k
}
\description{
Title
}
