% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surface_functions.R
\name{gamma_function_factory}
\alias{gamma_function_factory}
\title{Ganna function factory}
\usage{
gamma_function_factory(
  mode = c("smooth", "rotate", "identity"),
  au = 0.3,
  av = 0.25,
  angle = 0.5 * pi
)
}
\arguments{
\item{mode}{mode choice}

\item{au}{default 0.3}

\item{av}{default 0.25}

\item{angle}{}
}
\value{
a chosen gamma function
}
\description{
Ganna function factory
}
