% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{make_dphi_fn}
\alias{make_dphi_fn}
\title{Title}
\usage{
make_dphi_fn(basis, grad_f2k_fun, f2_fun, rotated = FALSE)
}
\arguments{
\item{basis}{}

\item{grad_f2k_fun}{}

\item{f2_fun}{}

\item{rotated}{}
}
\value{
a function (u,v) to 3x1 vector
}
\description{
Title
}
