% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize_gamma_scalar_input}
\alias{visualize_gamma_scalar_input}
\title{Gamma function plot in [0,1]^2}
\usage{
visualize_gamma_scalar_input(
  gamma_fun,
  n = 25,
  arrow_color = "tomato",
  point_color = "grey40",
  warp_scale = 1
)
}
\arguments{
\item{warp_scale}{}
}
\description{
Gamma function plot in [0,1]^2
}
